+++ .gitignore
@@ -71,5 +71,3 @@ fastlane/report.xml
 fastlane/Preview.html
 fastlane/screenshots
 fastlane/test_output
-
-.swiftlint.yml
