+++ .swiftlint.yml
@@ -1,0 +1,41 @@
+disabled_rules:
+- trailing_whitespace
+
+opt_in_rules:
+- empty_count
+- empty_string
+
+excluded:
+- Carthage
+- Pods
+- SwiftLint/Common/3rdPartyLib
+
+line_length:
+    warning: 150
+    error: 200
+    ignores_function_declarations: true
+    ignores_comments: true
+    ignores_urls: true
+    
+function_body_length:
+    warning: 300
+    error: 500
+
+function_parameter_count:
+    warning: 4
+    error: 6
+
+type_body_length:
+    warning: 300
+    error: 500
+
+file_length:
+    warning: 1000
+    error: 1500
+    ignore_comment_only_lines: true
+
+cyclomatic_complexity:
+    warning: 15
+    error: 25
+
+reporter: "xcode"
